[DI] Refactor unit tests (#4777)
Split inspected code into multiple files. This makes adding new tests
easier, as changes related to one inspected file doesn't influence
unrelated tests.
